move projects folder from js_modules and fix imports
