+++ src/common/components/Assets/Search.jsx
@@ -10,8 +10,8 @@ import { Formik, Form, Field } from 'formik';
 import { useRouter } from 'next/router';
 import { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
-import Icon from '../../common/components/Icon';
-import useDebounce from '../../common/hooks/useDebounce';
+import Icon from '../Icon';
+import useDebounce from '../../hooks/useDebounce';
 
 function Search({ placeholder, onChange }) {
   const router = useRouter();
