+++ src/common/components/Checkout/PaymentInfo.jsx
@@ -6,24 +6,24 @@ import {
 } from '@chakra-ui/react';
 import { useRouter } from 'next/router';
 import PropTypes from 'prop-types';
-import Heading from '../../common/components/Heading';
-import bc from '../../common/services/breathecode';
-import useSignup from '../../common/store/actions/signupAction';
+import Heading from '../Heading';
+import bc from '../../services/breathecode';
+import useSignup from '../../store/actions/signupAction';
 import 'react-datepicker/dist/react-datepicker.css';
-import useStyle from '../../common/hooks/useStyle';
-import useAuth from '../../common/hooks/useAuth';
-import { reportDatalayer } from '../../utils/requests';
-import { getQueryString, getStorageItem, getBrowserInfo } from '../../utils';
-import useCohortHandler from '../../common/hooks/useCohortHandler';
-import { getCohort } from '../../common/handlers/cohorts';
-import { getAllMySubscriptions } from '../../common/handlers/subscriptions';
-import { SILENT_CODE } from '../../lib/types';
+import useStyle from '../../hooks/useStyle';
+import useAuth from '../../hooks/useAuth';
+import { reportDatalayer } from '../../../utils/requests';
+import { getQueryString, getStorageItem, getBrowserInfo } from '../../../utils';
+import useCohortHandler from '../../hooks/useCohortHandler';
+import { getCohort } from '../../handlers/cohorts';
+import { getAllMySubscriptions } from '../../handlers/subscriptions';
+import { SILENT_CODE } from '../../../lib/types';
 import CardForm from './CardForm';
-import Icon from '../../common/components/Icon';
-import Text from '../../common/components/Text';
-import AcordionList from '../../common/components/AcordionList';
-import LoaderScreen from '../../common/components/LoaderScreen';
-import NextChakraLink from '../../common/components/NextChakraLink';
+import Icon from '../Icon';
+import Text from '../Text';
+import AcordionList from '../AcordionList';
+import LoaderScreen from '../LoaderScreen';
+import NextChakraLink from '../NextChakraLink';
 
 function PaymentInfo({ setShowPaymentDetails }) {
   const { t, lang } = useTranslation('signup');
