+++ src/common/components/Checkout/ServiceSummary.jsx
@@ -3,23 +3,23 @@ import { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import { Avatar, Box, Button, Link, Flex } from '@chakra-ui/react';
 import useTranslation from 'next-translate/useTranslation';
-import bc from '../../common/services/breathecode';
-import useSignup from '../../common/store/actions/signupAction';
-import Heading from '../../common/components/Heading';
-import Icon from '../../common/components/Icon';
-import useStyle from '../../common/hooks/useStyle';
-import useAuth from '../../common/hooks/useAuth';
-import Text from '../../common/components/Text';
-import AcordionList from '../../common/components/AcordionList';
-import NextChakraLink from '../../common/components/NextChakraLink';
-import LoaderScreen from '../../common/components/LoaderScreen';
-import { formatPrice, getStorageItem, getBrowserInfo } from '../../utils';
-import ModalInfo from '../../common/components/ModalInfo';
-import useCohortHandler from '../../common/hooks/useCohortHandler';
+import bc from '../../services/breathecode';
+import useSignup from '../../store/actions/signupAction';
+import Heading from '../Heading';
+import Icon from '../Icon';
+import Text from '../Text';
+import AcordionList from '../AcordionList';
+import NextChakraLink from '../NextChakraLink';
+import LoaderScreen from '../LoaderScreen';
+import ModalInfo from '../ModalInfo';
+import useStyle from '../../hooks/useStyle';
+import useAuth from '../../hooks/useAuth';
+import useCohortHandler from '../../hooks/useCohortHandler';
 import CardForm from './CardForm';
-import { SILENT_CODE } from '../../lib/types';
-import { reportDatalayer } from '../../utils/requests';
-import { BREATHECODE_HOST } from '../../utils/variables';
+import { SILENT_CODE } from '../../../lib/types';
+import { reportDatalayer } from '../../../utils/requests';
+import { BREATHECODE_HOST } from '../../../utils/variables';
+import { formatPrice, getStorageItem, getBrowserInfo } from '../../../utils';
 
 function ServiceSummary({ service }) {
   const { isAuthenticated } = useAuth();
