+++ src/pages/how-to/index.jsx
@@ -15,8 +15,8 @@ import PaginatedView from '../../common/components/PaginationView';
 import ProjectsLoader from '../../common/components/ProjectsLoader';
 import Text from '../../common/components/Text';
 import useFilter from '../../common/store/actions/filterAction';
-import Search from '../../js_modules/projects/Search';
-import TitleContent from '../../js_modules/projects/TitleContent';
+import Search from '../../common/components/Assets/Search';
+import TitleContent from '../../common/components/Assets/TitleContent';
 import { getQueryString } from '../../utils';
 import { ORIGIN_HOST, WHITE_LABEL_ACADEMY } from '../../utils/variables';
 import { parseQuerys } from '../../utils/url';
