+++ src/pages/how-to/technology/[technology].jsx
@@ -7,7 +7,7 @@ import Text from '../../../common/components/Text';
 import { toCapitalize } from '../../../utils';
 import { WHITE_LABEL_ACADEMY } from '../../../utils/variables';
 import Heading from '../../../common/components/Heading';
-import ProjectList from '../../../js_modules/projects/ProjectList';
+import ProjectList from '../../../common/components/Assets/ProjectList';
 import { parseQuerys } from '../../../utils/url';
 
 export const getStaticPaths = async ({ locales }) => {
