+++ src/pages/interactive-coding-tutorial/[slug].jsx
@@ -12,9 +12,9 @@ import Text from '../../common/components/Text';
 import Icon from '../../common/components/Icon';
 import useAuth from '../../common/hooks/useAuth';
 import useStyle from '../../common/hooks/useStyle';
-import FixedBottomCta from '../../js_modules/projects/FixedBottomCta';
-import SimpleTable from '../../js_modules/projects/SimpleTable';
-import TabletWithForm from '../../js_modules/projects/TabletWithForm';
+import FixedBottomCta from '../../common/components/Assets/FixedBottomCta';
+import SimpleTable from '../../common/components/Assets/SimpleTable';
+import TabletWithForm from '../../common/components/Assets/TabletWithForm';
 import ArticleMarkdown from '../../common/components/MarkDownParser/ArticleMarkdown';
 import { MDSkeleton } from '../../common/components/Skeleton';
 import getMarkDownContent from '../../common/components/MarkDownParser/markdown';
