+++ src/pages/interactive-exercise/[slug]/index.jsx
@@ -13,9 +13,9 @@ import getT from 'next-translate/getT';
 import Heading from '../../../common/components/Heading';
 import Link from '../../../common/components/NextChakraLink';
 import Text from '../../../common/components/Text';
-import TabletWithForm from '../../../js_modules/projects/TabletWithForm';
-import SimpleTable from '../../../js_modules/projects/SimpleTable';
-import FixedBottomCta from '../../../js_modules/projects/FixedBottomCta';
+import TabletWithForm from '../../../common/components/Assets/TabletWithForm';
+import SimpleTable from '../../../common/components/Assets/SimpleTable';
+import FixedBottomCta from '../../../common/components/Assets/FixedBottomCta';
 import TagCapsule from '../../../common/components/TagCapsule';
 import MarkDownParser from '../../../common/components/MarkDownParser';
 import useAuth from '../../../common/hooks/useAuth';
