+++ src/pages/interactive-exercises/index.jsx
@@ -9,9 +9,9 @@ import getT from 'next-translate/getT';
 import Text from '../../common/components/Text';
 import Icon from '../../common/components/Icon';
 import FilterModal from '../../common/components/FilterModal';
-import TitleContent from '../../js_modules/projects/TitleContent';
+import TitleContent from '../../common/components/Assets/TitleContent';
 import useFilter from '../../common/store/actions/filterAction';
-import Search from '../../js_modules/projects/Search';
+import Search from '../../common/components/Assets/Search';
 import { getQueryString } from '../../utils';
 import GridContainer from '../../common/components/GridContainer';
 import PaginatedView from '../../common/components/PaginationView';
