+++ src/pages/read/[slug].jsx
@@ -7,8 +7,8 @@ import { useRouter } from 'next/router';
 import Head from 'next/head';
 import Heading from '../../common/components/Heading';
 import Text from '../../common/components/Text';
-import Search from '../../js_modules/projects/Search';
-import TitleContent from '../../js_modules/projects/TitleContent';
+import Search from '../../common/components/Assets/Search';
+import TitleContent from '../../common/components/Assets/TitleContent';
 import Link from '../../common/components/NextChakraLink';
 import GridContainer from '../../common/components/GridContainer';
 import { cleanObject } from '../../utils';
